[CL-2110] Move auth provider
