+++ front/app/containers/Authentication/SignUpIn/SignUpInComponent/SignIn/index.tsx
@@ -2,7 +2,7 @@ import React, { memo, useCallback, useState, useEffect } from 'react';
 
 // components
 import PasswordSignin from './PasswordSignin';
-import AuthProviders, { AuthProvider } from '../AuthProviders';
+import AuthProviders, { AuthProvider } from 'components/AuthProviders';
 import Error from 'components/UI/Error';
 import {
   StyledHeaderContainer,
