+++ front/app/modules/commercial/id_vienna_saml/components/ViennaSamlButton.tsx
@@ -6,8 +6,8 @@ import { FormattedMessage } from 'utils/cl-intl';
 import messages from './messages';
 
 // components
-import { TOnContinueFunction } from 'components/SignUpIn/AuthProviderButton';
-import { StyledAuthProviderButton } from 'components/SignUpIn/AuthProviders';
+import { TOnContinueFunction } from 'components/AuthProviders/AuthProviderButton';
+import { StyledAuthProviderButton } from 'components/AuthProviders';
 import ViennaIcon from './ViennaIcon';
 
 // typings
